# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SyncStatus = typeof SyncStatus[keyof typeof SyncStatus];


export const SyncStatus = {
  pending: 'pending',
  synced: 'synced',
  failed: 'failed',
} as const;

export interface MemoCode {
  code: string;
  dateReceived: string;
}

export interface MemoFields {
  serviceMemoCode: string;
  dateReceived: string;
  dateOfServiceMemo: string;
  natureOfComplaint: string;
  accountNumber: string;
  consumerName: string;
  address: string;
  totalAmountPaid: string;
  orArNumber: string;
}

export type MemoExtraction = MemoFields & {
  /**
     * @minimum 0
     * @maximum 1
     */
  confidence: number;
  warnings: string[];
};

export type MemoExtractionInputImageMimeType = typeof MemoExtractionInputImageMimeType[keyof typeof MemoExtractionInputImageMimeType];


export const MemoExtractionInputImageMimeType = {
  'image/jpeg': 'image/jpeg',
  'image/webp': 'image/webp',
  'image/png': 'image/png',
} as const;

export interface MemoExtractionInput {
  /** Base64-encoded processed image */
  imageData: string;
  imageMimeType: MemoExtractionInputImageMimeType;
  serviceMemoCode: string;
  dateReceived: string;
}

export type MemoSubmissionImageMimeType = typeof MemoSubmissionImageMimeType[keyof typeof MemoSubmissionImageMimeType];


export const MemoSubmissionImageMimeType = {
  'image/jpeg': 'image/jpeg',
  'image/webp': 'image/webp',
  'image/png': 'image/png',
} as const;

export type MemoSubmission = MemoFields & {
  imageData: string;
  imageMimeType: MemoSubmissionImageMimeType;
};

export type ServiceMemo = MemoFields & ({
  id: number;
  driveStatus: SyncStatus;
  sheetStatus: SyncStatus;
  /** @nullable */
  driveImageLink?: string | null;
  /** @nullable */
  syncError?: string | null;
  createdAt: string;
});

export interface DashboardSummary {
  todayScans: number;
  totalScans: number;
  syncedCount: number;
  pendingCount: number;
  failedCount: number;
  recentMemos: ServiceMemo[];
}

export interface AppSettings {
  driveFolderId: string;
  spreadsheetId: string;
  sheetName: string;
  driveConnected: boolean;
  sheetsConnected: boolean;
}

export interface SettingsInput {
  driveFolderId: string;
  spreadsheetId: string;
  /** @minLength 1 */
  sheetName: string;
}

export type DriveItemKind = typeof DriveItemKind[keyof typeof DriveItemKind];


export const DriveItemKind = {
  folder: 'folder',
  spreadsheet: 'spreadsheet',
} as const;

export interface DriveItem {
  id: string;
  name: string;
  kind: DriveItemKind;
}

export interface DriveItemList {
  items: DriveItem[];
}

export interface SheetTabList {
  tabs: string[];
}

export type ListMemosParams = {
search?: string;
status?: ListMemosStatus;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};

export type ListMemosStatus = typeof ListMemosStatus[keyof typeof ListMemosStatus];


export const ListMemosStatus = {
  all: 'all',
  pending: 'pending',
  synced: 'synced',
  failed: 'failed',
} as const;

export type ListDriveItemsParams = {
parentId?: string;
type: ListDriveItemsType;
};

export type ListDriveItemsType = typeof ListDriveItemsType[keyof typeof ListDriveItemsType];


export const ListDriveItemsType = {
  folder: 'folder',
  spreadsheet: 'spreadsheet',
} as const;