# lib/api-zod/src/generated/types/appSettings.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AppSettings {
  driveFolderId: string;
  spreadsheetId: string;
  sheetName: string;
  driveConnected: boolean;
  sheetsConnected: boolean;
}