# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ServiceMemo } from './serviceMemo';

export interface DashboardSummary {
  todayScans: number;
  totalScans: number;
  syncedCount: number;
  pendingCount: number;
  failedCount: number;
  recentMemos: ServiceMemo[];
}