# lib/api-zod/src/generated/types/driveItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DriveItemKind } from './driveItemKind';

export interface DriveItem {
  id: string;
  name: string;
  kind: DriveItemKind;
}