# lib/api-zod/src/generated/types/driveItemList.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DriveItem } from './driveItem';

export interface DriveItemList {
  items: DriveItem[];
}