# lib/api-zod/src/generated/types/memoExtraction.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MemoFields } from './memoFields';

export type MemoExtraction = MemoFields & {
  /**
     * @minimum 0
     * @maximum 1
     */
  confidence: number;
  warnings: string[];
};