# lib/api-zod/src/generated/types/memoExtractionInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MemoExtractionInputImageMimeType } from './memoExtractionInputImageMimeType';

export interface MemoExtractionInput {
  /** Base64-encoded processed image */
  imageData: string;
  imageMimeType: MemoExtractionInputImageMimeType;
  serviceMemoCode: string;
  dateReceived: Date;
}