# lib/api-zod/src/generated/types/memoSubmission.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MemoFields } from './memoFields';
import type { MemoSubmissionImageMimeType } from './memoSubmissionImageMimeType';

export type MemoSubmission = MemoFields & {
  imageData: string;
  imageMimeType: MemoSubmissionImageMimeType;
};